requirements.txt
absl-py==2.1.0
albumentations==1.3.1
cachetools==5.3.3
captum==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dlib @ wheels/dlib-19.24.0-cp39-cp39-win_amd64.whl
efficientnet-pytorch==0.7.1
filelock==3.14.0
fonttools==4.51.0
google-auth==2.29.0
google-auth-oauthlib==0.4.6
grpcio==1.63.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.6.12
lazy_loader==0.4
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
munch==4.0.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pretrainedmodels==0.7.4
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qudida==0.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
tifffile==2024.5.3
torch @ https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp39-cp39-win_amd64.whl#sha256=245d04e1541350dba11c7b76e343ca0071bbcb10f956a09b6bede3d68db9e759
torchgeometry==0.1.2
torchmetrics==1.3.2
torchvision @ https://download.pytorch.org/whl/cu117/torchvision-0.15.2%2Bcu117-cp39-cp39-win_amd64.whl#sha256=8e0701a162cca13cee637397d7a67248bfeec35086cc753a1c513b2ea1dd476e
tqdm==4.66.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
zipp==3.18.1
